backend/requirements.txt
APScheduler==3.10.4
asyncpg==0.29.0
black==24.8.0
fastapi==0.112.2
jinja2
psycopg2-binary==2.9.9
pydantic-settings==2.4.0
pydantic==2.8.2
python-dotenv==1.0.1
ruff==0.6.4
SQLAlchemy-Utils==0.41.2
SQLAlchemy==2.0.34
slowapi[redis]==0.1.9
uvicorn==0.30.6